Yet another error... yesterday went well!
I got this working on VSC but then on the website it needed the answer as an integer, I changed it to an integer on the website but forgot that I had been testing and editing on VSC. Forgot to go back and fully change VSC to the original. The joys of working between two programs!